backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pandas
pdfplumber
scikit-learn
python-multipart
pydantic-settings
